backend_app/src/requirements.txt
backoff==2.2.1
Django==4.2.5
djangorestframework==3.15.2
django-split-settings==1.2.0
psycopg2-binary==2.9.9
django-cors-headers